requirements.txt
asgiref==3.2.3
attrs==19.3.0
Automat==20.2.0
cffi==1.14.0
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.0.3
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
lxml==4.5.0
parsel==1.5.2
Protego==0.1.16
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-decouple==3.3
pytz==2019.3
queuelib==1.5.0
Scrapy==2.0.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==19.10.0
w3lib==1.21.0
whitenoise==5.1.0
zope.interface==4.7.1
